cortex/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
Pillow==10.4.0
